requirements.txt
Django==4.2.7
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==4.2.7
django-browser-reload==1.12.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-sendgrid-v5==1.3.0
django-tailwind==3.6.0
django-unfold==0.71.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.10
inflection==0.5.1
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sendgrid==6.12.5
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
